+++ cspell.json
@@ -4,7 +4,9 @@
   "dictionaryDefinitions": [],
   "dictionaries": [],
   "words": [
+    "COLORCHIP",
     "flexdirection",
+    "IMAGECHIP",
     "Noto",
     "Pretendard"
   ],
