+++ src/pages/CreatePaperPage/CreatePaperPage.style.js
