+++ src/pages/MessageListPage/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MessageListPage";
