--- src/pages/PostMessagePage/index.js
@@ -1,3 +1,0 @@
-import PostMessagePage from "./PostMessagePage";
-
-export default PostMessagePage;
